addressing fit_ctd, flagging
